% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_and_graphics_functions.R
\name{autocorr}
\alias{autocorr}
\title{autocorr}
\usage{
autocorr(data, plot = F, q = 1)
}
\arguments{
\item{data}{a list containing:
\describe{
\item{Q}{an integer, the number of covariates,}
\item{x}{a list of matrices, the qth matrix contains the observation of the
      qth functional covariate at time points given by grids,}
\item{y}{a numerical vector, the outcome values y_i.}
}}

\item{plot}{a logical value. If it is TRUE, an image (heat map) of the
autocorrelation matrix is plotted. (optional)}

\item{q}{XXXX}
}
\value{
A symmetric matrix.
}
\description{
Compute the autocorrelation of the functional covariate sample.
}
\examples{
library(RColorBrewer)
### Autocorrelation of the function x_i(t)
param <- list(n=50,p=100,beta_type="smooth",Q=1)
data <- sim(param)
res_autocorr <- autocorr(data)
cols <- rev(colorRampPalette(brewer.pal(9,"YlOrRd"))(50))
image(res_autocorr,col=cols)
### Autocorrelation of the function beta_j(t).
data(data1)
data(param1)
# result of res_bliss1<-fit_Bliss(data=data1,param=param1)
data(res_bliss1)
\donttest{
# Example to modify!!
beta_functions_autocorr <- autocorr(list(grid = data1$grids[[1]],
                                         x = res_bliss1[[1]]))
image(beta_functions_autocorr)
}
}
