% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bliss_Gibbs_Sampler_multiple.R
\name{Bliss_Gibbs_Sampler_multiple}
\alias{Bliss_Gibbs_Sampler_multiple}
\title{Bliss_Gibbs_Sampler_multiple}
\usage{
Bliss_Gibbs_Sampler_multiple(data, param)
}
\arguments{
\item{data}{a list containing
\describe{
 \item{x_mult}{a list countaining the functions x_qi(t) observed at grids of time points}
 \item{y}{the outcome values y_i}
 \item{grids}{a list countaining the observation time points of the covariates functions.}
}}

\item{param}{a list which have to contain :
\describe{
\item{iter}{an integer, the number of iterations of the Gibbs sampler algorithm.}
\item{grids}{a list of numerical vectors, the qth vector is the grid of observation time points for the qth covariate.}
\item{K}{a vector of integers, hyperparameters of the Bliss model, the number of intervals for the Q covariates.}
\item{l_max}{a vector of integers, hyperparameters of the Bliss model. }
\item{basis}{a vector of characters among : "uniform" (default),
                "epanechnikov", "gauss" and "triangular". This indicates the
                shapes of the Q coefficient functions on the intervals. (optional)}
\item{eta_tilde}{a numerical vector of length (1+sum(K)), hyperparameter of the Bliss model.
                By default, eta is the vector (0,...,0). (optional)}
\item{V_tilde}{a matrix of dimension (1+sum(K))*(1+sum(K)), hyperparameter of the Bliss model. (optional)}
\item{a}{a nonnegative value, hyperparameter of the Bliss model. By default, a = 0.1. (optional)}
\item{b}{a nonnegative value, hyperparameter of the Bliss model. By default, b = 0.1. (optional)}
\item{g}{hyperparameter of the Bliss model,  a nonnegative value, the coefficient of the Zellner prior.}
\item{phi_m}{a list of numerical vectors. The priors of the mq, q=1,...,Q. If not specified, a uniform
                distribution is used for phi_m[[q]].}
\item{phi_l}{a list of numerical vectors. The priors of the lq, q=1,...,Q. If not specified, a uniform
                distribution is used for phi_l[[q]].}
\item{phi_l_mean}{a Q vector of numerical values. if "phi_l[[q]]" is "Gamma", phi_l_mean[q]
              corresponds to the mean of the Gamma prior of lq.}
\item{phi_l_sd}{a Q vector of numerical values. if "phi_l[[q]]" is "Gamma", phi_l_mean[q]
               corresponds to the standard deviation of the Gamma prior of lq.}
\item{prior_beta}{a character string, which indicates the prior on the
               beta_star[q]. The possible values are :
               1) "diag" (default) for a diagonal matrix prior,
               2) "Ridge_Zellner" for the Ridge Zellner prior}
\item{lambda}{a numerical value relative to the Ridge Zellner prior.}
\item{display}{a logical value. If FALSE nothing is printed.}
}}
}
\value{
a list containing :
\describe{
 \item{trace}{a matrix. Each row is an iteration of the Gibbs Sampler.}
 \item{param}{a list containing a, b, V_tilde, K, eta_tilde, l_max, grids and scale_ml}
}
}
\description{
Gibbs Sampler to sample from the posterior distribution of the Bliss model.
}
\examples{
data(data1)
data(param1)
res_Bliss_Gibbs_Sampler <- Bliss_Gibbs_Sampler_multiple(data1,param1)
K       <- param1$K
theta_1 <- res_Bliss_Gibbs_Sampler$trace[1,]
theta_1
}
