% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bliss_multiple.R
\name{Bliss_multiple}
\alias{Bliss_multiple}
\title{Bliss_multiple}
\usage{
Bliss_multiple(data, param, plot = FALSE, density = FALSE)
}
\arguments{
\item{data}{a list containing
\describe{
\item{1)}{the number of covariates}
\item{2)}{the functions x_qi(t) observed at grids of time points}
\item{3)}{the outcome values y_i}
}}

\item{param}{a list
\describe{
\item{iter}{an integer, the number of iterations of the Gibbs sampler algorithm.}
\item{grids}{a list of numerical vectors, the qth vector is the grid of observation points for the qth covariate.}
\item{K}{a vector of integers, hyperparameters of the Bliss model, the number of intervals for the Q covariates.}
\item{l_max}{a vector of integers, hyperparameters of the Bliss model. Beware, give the index
             corresponding to the value of the maximum width of the intervals (optional)}
\item{basis}{a vector of characters among : "uniform" (default),
                "epanechnikov", "gauss" and "triangular". This indicates the
                shapes of the Q coefficient functions on the intervals. (optional)}
\item{eta_tilde}{a numerical vector of length (1+sum(K)), hyperparameter of the Bliss model.
                By default, eta is the vector (0,...,0). (optional)}
\item{V_tilde}{a matrix of dimension (1+sum(K))*(1+sum(K)), hyperparameter of the Bliss model. (optional) (nasty code)}
\item{a}{a nonnegative value, hyperparameter of the Bliss model. Default, a = 0.1. (optional)}
\item{b}{a nonnegative value, hyperparameter of the Bliss model. Default, b = 0.1. (optional)}
\item{g}{hyperparameter of the Bliss model,  a nonnegative value,
               the coefficient of the Zellner prior.}
\item{phi_m}{a list of numerical vectors. The priors of the mq, q=1,...,Q. If not specified, a uniform
                distribution is used for phi_m[[q]].}
\item{phi_l}{a list of numerical vectors and/or characters. The priors of the lq, q=1,...,Q. If not specified, a uniform
                distribution is used for phi_l[[q]].}
\item{phi_l_mean}{a Q vector of numerical values. if "phi_l[[q]]" is "Gamma", phi_l_mean[q]
               corresponds to the mean of the Gamma prior of lq.}
\item{phi_l_sd}{a Q vector of numerical values. if "phi_l[[q]]" is "Gamma", phi_l_mean[q]
               corresponds to the standard deviation of the Gamma prior of lq.}
\item{prior_beta}{a character string, which indicates the prior on the
               beta_star[q]. The possible values are :
               1) "diag" (default) for a diagonal matrix prior,
               2) "Ridge_Zellner" for the Ridge Zellner prior}
\item{burnin}{an integer, the number of iteration to drop of the Gibbs sampler. (optional)}
\item{thin}{an integer, used to thin the Gibbs sample to compute an
                estimate of the posterior density of beta(t). (optional)}
\item{lims.kde}{a Qx2 matrix. lims.kde[q,] corresponds to the parameter (yl,yu) for the representation
                of the posterior density of the qth covariable. There are limits of the y-axis for the function kde2d. (optional)}
\item{n}{an integer, the number of grid points in each direction, for the kde2d function.
                See function kde2d. (optional)}
\item{iter_sann}{an integer, the number of iteration of the Simulated Annealing. (optional)}
\item{Temp}{a vector of nonnegative values, the Q initial temperatures for the
                cooling function of the Q Simulated Annealings. (optional)}
\item{k_max}{a vector of integers, k_max[q] is the maximum number of intervals for the
                function beta_q(t) at each iteration. (optional)}
\item{cols}{a vector of colors for the function image. Only if plot=TRUE. (optional)}
\item{new_grids}{a list of Q numerical vectors. If new_grids is not NULL, the
                coefficient functions beta_q(t) at each iteration are computed on these grids
                (only) to plot a graphical representation of the posterior distribution.}
\item{h1}{a vector of numerical values which are the Q bandwidths of the kernel
                density estimation for the t-axis (optional). h1[q] the bandwidth of the kernel density estimation for the qth covariate.}
\item{ylim}{a Qx2 matrix, the qth line gives the limits for the y-axis for the plotting function
                "image_Bliss" for the qth covariate.}
\item{main}{a vector of characters, main[q] is for the plotting function
                "image_Bliss" for the qth covariate.}
\item{n_chains}{number of chains to do in the Gibbs sampler.}
}}

\item{plot}{a logical value. If TRUE, the Bliss estimates and
the estimates of the posterior density are plotted. (optional)}

\item{density}{a logical value. If TRUE, the posterior density is estimated. (optional)}
}
\value{
return a list containing :
\describe{
 \item{Bliss_estimate}{a list of numerical vectors, the Bliss estimates : we obtain Q estimates, one for each covariable.}
 \item{posterior_density_estimate}{a list of lists containing the estimates of the posterior density.
           (obtained with the kde2d function). Firts level of the list: Q lists, for the Q covariates. Second
           level: for each covariate we have a list of three components, that is the  x and y coordinates of the grid points, and z
           a matrix of the estimated density: rows correspond to the value of x, columns to the value of y.}
 \item{beta_functions}{a list of matrices. For the qth covariate, beta_functions[[q]] is a matrix where each row is a
                function beta_qi associated to the iteration i of the Gibbs sampler.}
 \item{res.Simulated_Annealing}{a list of lists: res.Simulated_Annealing[[q]] is the result of the
                function Bliss_Simulated_Annealing applied for the qth covariate.}
 \item{res.Gibbs_Sampler}{a list of lists: res.Gibbs_Sampler[[q]] is the the result of the function
                Bliss_Gibbs_Sampler for the qth covariate.}
}
}
\description{
perform the Bliss method to obtain an estimate
              of the Q coefficient functions of the functional linear
              regression model with Q functional covariates.
}
\examples{
# see the vignette BlissIntro.
}
