% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bliss_Simulated_Annealing.R
\name{Bliss_Simulated_Annealing}
\alias{Bliss_Simulated_Annealing}
\title{Bliss_Simulated_Annealing}
\usage{
Bliss_Simulated_Annealing(beta_functions, param, scale_ml)
}
\arguments{
\item{beta_functions}{a matrix. Each row is a coefficient function computed from the posterior sample.}

\item{param}{a list containing :
\describe{
\item{grid}{a numerical vector which countains the observation time points.}
\item{burnin}{an integer, the number of iterations to drop from the Gibbs sample. (optional)}
\item{iter}{an integer, the number of iteration of the Gibbs Sampler algorithm.}
\item{iter_sann}{an integer, the number of iteration of the Simulated Annealing algorithm. (optional)}
\item{Temp}{a non negative value, the initial temperature for the cooling function of the Simulated Annealing. (optional)}
\item{k_max}{an integer, the maximum number of intervals. (optional)}
\item{l_max_sann}{an integer, the maximum value for the parameter l. (optional)}
\item{basis}{a character vectors, used to compute the coefficient function, see the function beta_build. (optional)}
}}

\item{scale_ml}{a matrix returned by the function "Bliss_Gibbs_Sampler".}
}
\value{
a list containing :
\describe{
 \item{Bliss_estimate}{a numerical vector, corresponding to the Bliss estimate.}
 \item{posterior_expe}{a numerical vector, which is the posterior expectation of beta(t), for each t in the
                grid ot time points.}
 \item{posterior_var}{a numerical vector, which is the posterior expectation of beta(t), for each t in the
                grid ot time points.}
 \item{trace}{Each row is an iteration of the Simulated Annealing algorithm.}
 \item{argmin}{an integer, which is the index of the iteration minimizing the loss.}
}
}
\description{
Simulated Annealing algorithm to determine the minimum of the posterior expectation Bliss loss, i.e. the
              Bliss estimate.
}
\examples{
data(data1)
data(param1)
#res.Simulated_Annealing <- Bliss_Simulated_Annealing(beta_functions,param1)
#ylim <- range(c(res.Simulated_Annealing$Bliss_estimate,
 #                res.Simulated_Annealing$posterior_expe))
#plot(param$grid,res.Simulated_Annealing$Bliss_estimate,type="l",ylim=ylim)
#lines(param$grid,res.Simulated_Annealing$posterior_expe,lty=2)
}
