% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bliss_Simulate_Data.R
\name{choose_beta}
\alias{choose_beta}
\title{choose_beta}
\usage{
choose_beta(param)
}
\arguments{
\item{param}{a list containing
\describe{
 \item{p}{a numerical value}
 \item{grid}{a numerical vector}
 \item{beta_type}{a character vector to choose between "smooth", "random_smooth", "simple", "simple2", "random_simple",
                  "sinusoid", "flat_sinusoid", "sharp"}
}}
}
\value{
a numerical vector which is the observation of the coefficient function at the given grid of times points: "grid".
}
\description{
compute a coefficient function for the function linear model.
}
\details{
several shapes are available for the coefficient function.
}
\examples{
### smooth
param <- list(p=100,grid=seq(0,1,length=100),beta_type="smooth")
beta_function <- choose_beta(param)
plot(param$grid,beta_function,type="l")
### random_smooth
param <- list(p=100,grid=seq(0,1,length=100),beta_type="random_smooth")
beta_function <- choose_beta(param)
plot(param$grid,beta_function,type="l")
### simple
param <- list(p=100,grid=seq(0,1,length=100),beta_type="simple")
beta_function <- choose_beta(param)
plot(param$grid,beta_function,type="s")
### simple2
param <- list(p=100,grid=seq(0,1,length=100),beta_type="simple2")
beta_function <- choose_beta(param)
plot(param$grid,beta_function,type="s")
### random_simple
param <- list(p=100,grid=seq(0,1,length=100),beta_type="random_simple")
beta_function <- choose_beta(param)
plot(param$grid,beta_function,type="s")
### sinusoid
param <- list(p=100,grid=seq(0,1,length=100),beta_type="sinusoid")
beta_function <- choose_beta(param)
plot(param$grid,beta_function,type="l")
### flat_sinusoid
param <- list(p=100,grid=seq(0,1,length=100),beta_type="flat_sinusoid")
beta_function <- choose_beta(param)
plot(param$grid,beta_function,type="l")
### sharp
param <- list(p=100,grid=seq(0,1,length=100),beta_type="sharp")
beta_function <- choose_beta(param)
plot(param$grid,beta_function,type="l")
}
