% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{compute_beta_functions_mult}
\alias{compute_beta_functions_mult}
\title{compute_beta_functions}
\usage{
compute_beta_functions_mult(res.Gibbs_Sampler, param)
}
\arguments{
\item{res.Gibbs_Sampler}{a list (provided by the function Bliss_Gibbs_Sampler).}

\item{param}{(optional) a list containing
\describe{
\item{p}{the number of time points}
\item{K}{the number of intervals}
\item{grid}{the grid of time points}
\item{basis}{a character vector}
}}
}
\value{
return a matrix. Each row is a function observed on the grid of time points.
}
\description{
Compute the function beta_i(t) for each iteration i of the Gibbs sample.
}
\examples{
library(RColorBrewer)
data(data1)
data(param1)
res_Bliss_mult <- Bliss_multiple(data=data1,param=param1)
beta_functions <- compute_beta_functions_mult(res_Bliss_mult$res.Gibbs_Sampler,param1)
indexes <- sample(nrow(beta_functions[[1]]),1e2,replace=FALSE)
cols <- colorRampPalette(brewer.pal(9,"YlOrRd"))(1e2)
matplot(param1$grids[[1]],t(beta_functions[[1]][indexes,]),type="l",lty=1,col=cols)
}
