% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{compute_beta_sample}
\alias{compute_beta_sample}
\title{compute_beta_sample}
\usage{
compute_beta_sample(posterior_sample, param, Q, progress = FALSE)
}
\arguments{
\item{posterior_sample}{a list provided by the function Bliss_Gibbs_Sampler.}

\item{param}{a list containing: (optional)
\describe{
\item{p}{XXXXXX}
\item{K}{a vector of integers, corresponding to the numbers of intervals for
      each covariate.}
\item{grids}{a numerical vector, the observation time points.}
\item{basis}{a vector of characters among : "uniform" (default),
      "epanechnikov", "gauss" and "triangular" which correspond to
      different basis functions to expand the coefficient function and the
      functional covariates (optional)}
}}

\item{Q}{numeric}

\item{progress}{a logical value. If TRUE, the algorithm progress is displayed.
(optional)}
}
\value{
return a matrix. Each row is a function evaluated on the grid of
        time points.
}
\description{
Compute the coefficient function for each iteration of the
             Gibbs sample.
}
\examples{
\donttest{
library(RColorBrewer)
data(data1)
data(param1)
res_Bliss_mult <- fit_Bliss(data=data1,param=param1)
beta_sample <- compute_beta_sample_mult(res_Bliss_mult$res_Gibbs_Sampler,param1)
indexes <- sample(nrow(beta_sample[[1]]),1e2,replace=FALSE)
cols <- colorRampPalette(brewer.pal(9,"YlOrRd"))(1e2)
matplot(param1$grids[[1]],t(beta_sample[[1]][indexes,]),type="l",lty=1,col=cols)
}
}
