% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_and_graphics_functions.R
\name{compute_chains_info}
\alias{compute_chains_info}
\title{compute_chains_info
compute_chains_info}
\usage{
compute_chains_info(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
compute_chains_info
compute_chains_info
}
\examples{
# compute x
}
