% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Data.R
\name{compute_random_walk}
\alias{compute_random_walk}
\title{compute_random_walk}
\usage{
compute_random_walk(n, p, mu, sigma, start = rep(0, n))
}
\arguments{
\item{n}{an integer, the number of random walk.}

\item{p}{an integer, the length of the random walks.}

\item{mu}{a numerical vector, the average random walk.}

\item{sigma}{a numerical value which is the standard deviation of the
gaussian distribution used to compute the random walk.}

\item{start}{a numerical vector which is the initial value of
the random walks. (optional)}
}
\value{
a matrix where each row is a random walk.
}
\description{
compute a random walk. (gaussian)
}
\details{
see the \code{\link[=sim_x]{sim_x}} function.
}
\examples{
# see the sim_x() function.
}
