% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{compute_starting_point}
\alias{compute_starting_point}
\title{compute_starting_point}
\usage{
compute_starting_point(beta_expe)
}
\arguments{
\item{beta_expe}{numeric}
}
\value{
a value
}
\description{
compute_starting_point
}
\examples{
data(data1)
data(param1)
}
