% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{compute_starting_point_sann}
\alias{compute_starting_point_sann}
\title{compute_starting_point_sann}
\usage{
compute_starting_point_sann(beta_expe)
}
\arguments{
\item{beta_expe}{a numerical vector, the expectation of the coefficient
function posterior sample.}
}
\value{
a matrix with 3 columns : "m", "l" and "b". The two first
        columns define the begin and the end of the intervals and the third
        gives the mean values of each interval.
}
\description{
Compute a starting point for the Simulated Annealing algorithm.
}
\examples{
data(res_bliss1)
mystart<-compute_starting_point_sann(apply(res_bliss1$beta_sample[[1]],2,mean))
}
