% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Data.R
\name{corr_matrix}
\alias{corr_matrix}
\title{corr_matrix}
\usage{
corr_matrix(diagonal, ksi)
}
\arguments{
\item{diagonal}{a numerical vector corresponding to the diagonal of the final matrix.}

\item{ksi}{a "coefficient of correlation". See the article Bliss, Section 3.1 for more details.}
}
\value{
a symmetric matrix.
}
\description{
compute an autocorrelation matrix to simulate functions x_i(t).
}
\examples{
### Test 1 : weak autocorrelation
ksi     <- 1
diagVar <- abs(rnorm(100,50,5))
Sigma   <- corr_matrix(diagVar,ksi^2)
persp(Sigma)
### Test 2 : strong autocorrelation
ksi     <- 0.2
diagVar <- abs(rnorm(100,50,5))
Sigma   <- corr_matrix(diagVar,ksi^2)
persp(Sigma)
}
