% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data1}
\alias{data1}
\title{a list of data}
\format{a list of data
\describe{
  \item{Q}{the number of functional covariate}
  \item{y}{y coordinate}
  \item{x}{x coordinate}
  \item{betas}{}
  \item{grids}{The grid of the time observations}
}}
\usage{
data1
}
\description{
A data object for bliss model
}
\keyword{datasets}
