% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{diagnostics}
\alias{diagnostics}
\title{diagnostics}
\usage{
diagnostics(chains, param)
}
\arguments{
\item{chains}{a list
\describe{
\item{res.Gibbs_Sampler}{a list resulting of the function Bliss_Gibbs_Sampler.}
\item{beta_functions}{a matrix which is the result of the function compute_beta_functions.}
\item{posterior_density_estimate}{a list which is the result of the function density_estimation.}
}}

\item{param}{a list
\describe{
\item{iter}{the number of iterations.}
\item{burnin}{the number of iterations to drop.}
\item{K}{the number of intervals of the beta_i's functions.}
\item{p}{the number of time points.}
\item{ts}{a vector. The sample of time points t_j used for compute some diagnostics of beta(t_j). (optional)}
\item{l_ts}{an integer, the number of time points to sample, if ts is not specified. (optional)}
\item{lag_max}{an integer, the maximal lag when compute the autocorrelation of the trace.}
}}
}
\value{
a list containing the diagnostics which can be plotted with the
        function "plot_diagnostics".
}
\description{
perform some diagnostics for the chains resulting of the
             Gibbs Sampler algorithm.
}
\details{
The implementation does not tackle the issue of several functional covariates.
}
\examples{
data(data1)
data(param1)
param1$n_chains <- 3
param1$iter <- 1e3
param1$burnin <- 1e2
param1$display <- FALSE
param1$compute_posterior <- FALSE
res_bliss_chains <- Bliss_multiple(data1,param1)
res_diagnostic <- diagnostics(res_bliss_chains$chains,param1)
cols <- makeTransparent(2:5,0.8)
plot(res_diagnostic$hist_mu[[1]],border=0,col=cols[1],main="",xlab="",ylab="")
for(c in 2:3){
   plot(res_diagnostic$hist_mu[[c]],add=TRUE,border=0,col=cols[c])
}
plot(res_diagnostic$density_beta[[4]][[1]],main="",xlab="",ylab="")
for(c in 2:3){
   lines(res_diagnostic$density_beta[[4]][[c]],col=c)
}
}
