% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{do_need_to_reduce}
\alias{do_need_to_reduce}
\title{do_need_to_reduce}
\usage{
do_need_to_reduce(param)
}
\arguments{
\item{param}{a list containing  p_threshold the maximum number of time points
and p the actual number of time points for each functional covariate.}
}
\value{
a boolean value.
}
\description{
Determine if it is required to reduce the size of the grid time
       points for each functional covariate.
}
\examples{
data(param1)
param1$p <- sapply(data1$grids,length)

do_need_to_reduce(param1)
}
