% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{dposterior}
\alias{dposterior}
\title{dposterior}
\usage{
dposterior(res.Gibbs_Sampler, data, theta = NULL)
}
\arguments{
\item{res.Gibbs_Sampler}{a list corresponding to the object returns by the "Bliss_Gibbs_Sampler_multiple" function.}

\item{data}{a list containing
\describe{
\item{1)}{the number of covariates,}
\item{2)}{the functions x_qi(t) observed at grids of time points,}
\item{3)}{the outcome values y_i.}
}}

\item{theta}{a matrix or a vector which contains the parameter value(s) for which the posterior density is claimed.}
}
\value{
return the posterior density and the posterior log density for a given parameters set.
}
\description{
Compute the posterior distribution of the model parameters
}
\details{
If the option theta is null, the posterior density is computed for the MCMC sample given in the "res.Gibbs_Sampler" object.
}
\examples{
data(data1)
data(param1)
res_Bliss_mult <- Bliss_multiple(data=data1,param=param1)
# Compute the posterior density of the MCMC sample :
res_poste <- dposterior(res_Bliss_mult$res.Gibbs_Sampler,data1)
### Compute the posterior density of the MCMC sample :
theta <- list( beta= rbind(c(3,4,-1), c(3,2,1)),
m = rbind(c(10,50,80),c(50,10,10)),
l = rbind(c(10,10,5),c(10,5,5)),
mu=c(1,1.5),sigma_sq=c(1,0.5))
dposterior(res_Bliss_mult$res.Gibbs_Sampler,data1,theta=theta)
}
