% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{finer_grid}
\alias{finer_grid}
\title{finer_grid}
\usage{
finer_grid(curve, grid, new_grid)
}
\arguments{
\item{curve}{a numerical vector.}

\item{grid}{a numerical vector, the former grid.}

\item{new_grid}{a numerical vector.}
}
\value{
a numerical vector, the curve evaluated on the new grid.
}
\description{
Compute a curve on a new finer gird.
}
\details{
This is nasty code.
}
\examples{
grid <- seq(0,1,l=1e1)
new_grid <- seq(0,1,l=1e2)
curve <- 3*grid^2 + sin(grid*2*pi)
plot(grid,curve,type="o")
lines(new_grid,finer_grid(curve,grid,new_grid),type="o",col=makeTransparent(2))
}
