% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{Fourier_basis_build}
\alias{Fourier_basis_build}
\title{Fourier_basis_build}
\usage{
Fourier_basis_build(grid, dim, per = 2 * pi)
}
\arguments{
\item{grid}{a numerical vector.}

\item{dim}{a numerical value. It corresponds to dim(basis)/2.}

\item{per}{a numerical value which corresponds to the period of the sine and cosine functions.}
}
\value{
a matrix.
}
\description{
define a Fourier basis to simulate functions x_i(t).
}
\details{
see the \code{\link[=sim_functions]{sim_functions}} function.
}
\examples{
# see the sim_functions() function.
}
