% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_and_graphics_functions.R
\name{interpretation_plot}
\alias{interpretation_plot}
\title{interpretation_plot}
\usage{
interpretation_plot(estimate, data, q = 1, centered = FALSE, cols = NULL)
}
\arguments{
\item{estimate}{a numerical vector, the Bliss estimate.}

\item{data}{a list containing:
\describe{
\item{Q}{an integer, the number of covariates,}
\item{x}{a list of matrices, the qth matrix contains the observation of the
      qth functional covariate at time points given by grids,}
\item{y}{a numerical vector, the outcome values y_i.}
}}

\item{q}{an integer, the index of the functional covariate to plot.}

\item{centered}{a logical value. If TRUE, the functional data are centered.}

\item{cols}{a numerical vector of colours.}
}
\description{
Provide a graphical representation of the functional data
             with a focus on the detected periods with the Bliss method.
}
\examples{
\donttest{
# Not run!!
data(data1)
data(param1)
res_bliss1 <- fit_Bliss(data=data1,param=param1,progress=TRUE)
data(res_bliss1)
interpretation_plot(res_bliss1$Bliss_estimate[[1]],data1)
interpretation_plot(res_bliss1$Bliss_estimate[[1]],data1,centered=FALSE)
}
}
