% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{lines_step_function}
\alias{lines_step_function}
\title{lines_step_function}
\usage{
lines_step_function(extended_grid, fct, bound = FALSE, ...)
}
\arguments{
\item{extended_grid}{a vector. Similar to the grid of the functional
covariate but it is adapted to provide suitable representation.}

\item{fct}{the numerical vector to plot.}

\item{bound}{a boolean value. If bound is TRUE, the plot of fct is
one line. Otherwise, several lines are used to plot fct.}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}
}
\description{
a suitable representation of the Bliss estimate of the coefficient function.
}
\examples{
### Plot the BLiss estimate on a suitable grid
}
