% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{makeTransparent}
\alias{makeTransparent}
\title{makeTransparent}
\usage{
makeTransparent(..., alpha = 0.5)
}
\arguments{
\item{...}{must be a color. Numerical or character string.}

\item{alpha}{a numerical value between 0 and 1, corresponding to the
transparency of the returned color.}
}
\value{
a character string coding for a color.
}
\description{
Make transparent some color.
}
\examples{
cols <- makeTransparent(2:6,0.3)
res_hist <- hist(rnorm(1e4,0,1),nclass=2e2,border = 0,col = cols[1],
                 xlim=c(-2,6),xlab="",main = "")
hist(rnorm(1e4,1,1),border = 0,col = cols[2],add=TRUE,nclass=2e2)
hist(rnorm(1e4,2,1),border = 0,col = cols[3],add=TRUE,nclass=2e2)
hist(rnorm(1e4,3,1),border = 0,col = cols[4],add=TRUE,nclass=2e2)
hist(rnorm(1e4,4,1),border = 0,col = cols[5],add=TRUE,nclass=2e2)
}
\seealso{
Thanks to Ricardo Oliveros-Ramos for the function on the web :
              \url{http://stackoverflow.com/questions/8047668/transparent-equivalent-of-given-color}
}
