% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{pdexp}
\alias{pdexp}
\title{pdexp}
\usage{
pdexp(a, l_values)
}
\arguments{
\item{a}{a positive value, the mean of the Exponential prior.}

\item{l_values}{a numerical value, the discrete support of the parameter l.}
}
\value{
a numerical vector, which is the prability function on "l_values".
}
\description{
Probability function of a discretized Exponentiel distribution.
Compute the probability function of the Exponential prior on l. #XXXXXXXX
}
\examples{
pdexp(10,seq(0,1,1))
}
