% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{pexp_grid}
\alias{pexp_grid}
\title{dexp_grid}
\usage{
pexp_grid(a, l_values)
}
\arguments{
\item{a}{a positive value, the mean of the Exponential prior.}

\item{l_values}{a numerical value, the discrete support of the parameter l.}
}
\value{
a numerical vector, which is the prability function on "l_values".
}
\description{
Compute the probability function of the Exponential prior on l. #XXXXXXXX
}
\examples{
# not run
}
