% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{plot_diagnostics}
\usage{
plot_diagnostics(res_diagnostics, param, chain = NULL, which_plot = NULL,
  time = NULL)
}
\arguments{
\item{res_diagnostics}{a list containing the result of \code{\link[=diagnostics]{diagnostics}} function.}

\item{param}{a list containing n_chains, the number of chains.}

\item{chain}{an integer, corresponding to the index of the chain to those diagnositcs have to be plotted.
 If chain is NULL (defaut), the diagnostics of all the chains are plotted on
the same plots. (optional)}

\item{which_plot}{a string character indicating which parameter is of interest. The possible value : 'mu', 'sigma_sq' or 'beta'.}

\item{time}{a numerical value belonging to the "ts" vector (option
of \code{\link[=diagnostics]{diagnostics}} function). (optional)}
}
\value{
a graph
}
\description{
: Plot the diagnostics computed by \code{\link[=diagnostics]{diagnostics}}
}
\examples{
data(data1)
data(param1)
param1$n_chains <- 3
param1$iter <- 1e3
param1$burnin <- 1e2
param1$display <- FALSE
param1$compute_posterior <- FALSE
res_bliss_chains <- Bliss_multiple(data1,param1)
res_diagnostic <- diagnostics(res_bliss_chains$chains,param1)
plot_diagnostics(res_diagnostic,param1,which_plot="mu")
}
