% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{plot_impact_mult}
\alias{plot_impact_mult}
\title{plot_impact_mult}
\usage{
plot_impact_mult(estimate, data, q = 1, centered = TRUE, cols = NULL)
}
\arguments{
\item{estimate}{a numerical vector, the estimate of the coefficient function.}

\item{data}{a list containing y, x and grid.}

\item{q}{an integer, the index of the functional covariate to consider.}

\item{centered}{a logical value. If TRUE, the curve x are centered.}

\item{cols}{a numerical vector of colors.}
}
\description{
provide a graphical representation of the data with a focus on the detected periods with the Bliss method.
}
\examples{
\donttest{
data(data1)
data(param1)
res_Bliss_mult <- Bliss_multiple(data=data1,param=param1)
plot_impact_mult(res_Bliss_mult$Bliss_estimate[[1]],data1)
plot_impact_mult(res_Bliss_mult$Bliss_estimate[[1]],data1,centered=FALSE)
}
}
