% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bliss_method.R
\name{printbliss}
\alias{printbliss}
\title{Print a bliss Object}
\usage{
printbliss(x, ...)
}
\arguments{
\item{x}{input bliss Object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print a bliss Object
}
\examples{
data(data1)
data(param1)
#res_Bliss_mult <- fit_Bliss(data=data1,param=param1)
# printbliss(res_Bliss_mult)
}
