% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{prior_l}
\alias{prior_l}
\title{prior_l}
\usage{
prior_l(m, s, grid_l)
}
\arguments{
\item{m}{a positive value, the mean of the Gamma prior.}

\item{s}{a nonnegative value, the standard deviation of the Gamma prior.}

\item{grid_l}{a numerical value, the discrete support of the parameter l.}
}
\value{
a numerical vector, which is the prability function on "grid_l".
}
\description{
Compute the probability function of the Gamma prior on l.
}
\examples{
grid_l <- seq(0,5,l=100)
f <- prior_l(3,1,grid_l)
plot(grid_l,f,type="h",xlab="",ylab="")
f <- prior_l(1,0.5,grid_l)
plot(grid_l,f,type="h",xlab="",ylab="")
f <- prior_l(1,5,grid_l)
plot(grid_l,f,type="h",xlab="",ylab="")
}
