% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{reduce_x}
\alias{reduce_x}
\title{reduce_x}
\usage{
reduce_x(data, param)
}
\arguments{
\item{data}{similar to \code{fit_Bliss}.}

\item{param}{a list containing values \code{Q}, \code{p} and \code{p}}
}
\value{
a numerical value, the approximation.
}
\description{
Reduce the number of time points.
}
\examples{
param <- list(Q=1,n=10,p=c(150),grids_lim=list(c(0,1)))
data <- sim(param)

data(param1)
param1$n <- nrow(data$x[[1]])
param1$p <- sapply(data$grids,length)
param1$Q <- length(data$x)

data <- reduce_x(data,param1)
}
