% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Data.R
\name{sigmoid_sharp}
\alias{sigmoid_sharp}
\title{sigmoid_sharp}
\usage{
sigmoid_sharp(x, loc = 0, ...)
}
\arguments{
\item{x}{a numerical vector, time points.}

\item{loc}{a numerical value (optional), the time of the sharp.}

\item{...}{Arguments (optional) for the function sigmoid.}
}
\value{
a numerical vector.
}
\description{
Compute a sharp sigmoid function.
}
\details{
see the function \code{\link[=sim_x]{sim_x}}.
}
\examples{
## Test 1 :
x <- seq(-7,7,0.1)
y <- sigmoid_sharp(x)
plot(x,y,type="l",main="Sharp sigmoid")
## Test 2 :
x  <- seq(-7,7,0.1)
y  <- sigmoid_sharp(x,loc=3)
y2 <- sigmoid_sharp(x,loc=3,asym=0.5)
y3 <- sigmoid_sharp(x,loc=3,v   =  5)
plot(x,y,type="l",main="Other sharp sigmoids")
lines(x,y2,col=2)
lines(x,y3,col=3)
}
