% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Data.R
\name{sim_correlated_functions}
\alias{sim_correlated_functions}
\title{sim_correlated_x}
\usage{
sim_correlated_functions(param)
}
\arguments{
\item{param}{a list containing :
\describe{
 \item{n}{an integer, the number of functions.}
 \item{p}{an integer, the number of observation times.}
 \item{grid}{a numerical vector, the grid of observation times.}
 \item{x_shape}{a character vector, the shape of the functions x_i(t). (optional)}
 \item{dim}{a numerical value, the dimension of the Fourier basis, if "shape" is "Fourier" or "Fourier2". (optional)}
 \item{ksi}{a numerical value, a "coefficient of correlation", see the Bliss article Section 3.1 for more details.}
 \item{diagVar}{a numerical vector, the diagonal of the autocorrelation matrix of the functions x_i(t).}
}}
}
\value{
todo
}
\description{
todo
}
\examples{
#todo
}
