% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bliss_Simulate_Data.R
\name{sim_multiple}
\alias{sim_multiple}
\title{sim_multiple}
\usage{
sim_multiple(param)
}
\arguments{
\item{param}{a list containing :
\describe{
 \item{n}{an integer, the number of observations.}
 \item{p}{a vector of integers, the qth component is the number of observation times for the qth covariate.}
 \item{b_inf}{a vector of numerical values, the qth component is the lower boundary of the support for the
                     observation times for the qth covariate. (optional)}
 \item{b_sup}{a vector of numerical values, the qth component is the upper boundary of the support for the
                     observation times for the qth covariate. (optional)}
 \item{mu}{a numerical value, the intercept of the model. (optional)}
 \item{r}{a nonnegative value, the signal to noise ratio. (optional)}
 \item{links}{a list of functions, the qth element of the list is a link function to simulate data from a GFLM. (optional)}
 \item{grids}{a list of numerical vectors, the qth vector is the grid of observation points for the qth covariate
                     (optional, can be deduced from b_inf,b_sup and p).}
 \item{beta_types}{a character vector. It indicates the shapes of
                    the coefficient functions. If we have Q covariates, Q shapes can be indicated.}
 \item{x_types}{a character vector. It indicates the shapes of the
                    functions x_qi(t). If we have Q covariates, Q shapes can be indicated. (optional)}
 \item{dim}{vector of numerical values, the dimensions of the Fourier basis, if "type" is "Fourier" or "Fourier2".
                     One value for each covariate. (optional); if the qth shape is not "Fourier" ou "Fourier2", the associated component should be specified as NA (optional)}
 \item{ksi}{vector of numerical values which are "coefficients of correlation" (one value for each covariate),
                    see the Bliss article Section 3.1 for more details. (optional)}
 \item{diagVar}{a list of numerical vectors, the qth vector is the diagonal of the
                    autocorrelation matrix of the functions x_qi(t). (optional)}
}}
}
\value{
a list containing :
\describe{
 \item{y}{a numerical vector, the outcomes y_i.}
 \item{x_mult}{a list of matrices, the qth matrice representing the qth functional covariate on a grid of
                times points (grid[[q]]).}
 \item{beta_function_mult}{a list of numerical vectors, the qth vector representing the coefficient function of
                the qth covariate on a grid of times points (grid[[q]]).}
 \item{grids}{a list of numerical vectors, the qth vector is the grid of observation points for the qth covariate.}
}
}
\description{
simulate a dataset for the functional linear model.
}
\examples{
library(RColorBrewer)
param <- list(n=10,p=c(200,100),beta_types=c("simple","smooth"),b_inf=c(0,-2.12),b_sup=c(1,3.14))
data <- sim_multiple(param)
data$y
data$expe
cols <- colorRampPalette(brewer.pal(9,"YlOrRd"))(10)
q=2
par(mfrow=c(2,1))
matplot(data$grids[[q]],t(data$x_mult[[q]]),type="l",lty=1,col=cols)
plot(data$grids[[q]],data$beta_function_mult[[q]],type="l")
abline(h=0,lty=2,col="gray")
par(mfrow=c(1,1))
}
