% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{support_estimation}
\alias{support_estimation}
\title{support_estimation}
\usage{
support_estimation(beta_functions)
}
\arguments{
\item{beta_functions}{a matrix for which each row is an evaluation of a function on a grid.}
}
\value{
a list
\describe{
\item{alpha_t}{a numerical vector. For each point t, alpha(t) is the probability that beta(t) is non-null}
\item{estimate}{a matrix given by the function "interval_detection", which give the estimation of
the support of the coefficient function}
}
}
\description{
Compute the estimate of the support of the coefficient function.
}
\examples{
data(data1)
data(param1)
res_Bliss_mult <- Bliss_multiple(data=data1,param=param1)
res.support <- support_estimation(res_Bliss_mult$beta_functions[[1]])
 ### The estimate
 res.support$estimate
 ### Plot the result
 grid <- res_Bliss_mult$param$grids[[1]]
 plot(grid,res.support$alpha_t,ylim=c(0,1),type="l",xlab="",ylab="")
 for(k in 1:nrow(res.support$estimate)){
    segments(grid[res.support$estimate[k,1]],0.5,
           grid[res.support$estimate[k,2]],0.5,lwd=2,col=2)
   points(grid[res.support$estimate[k,1]],0.5,pch="|",lwd=2,col=2)
   points(grid[res.support$estimate[k,2]],0.5,pch="|",lwd=2,col=2)
 }
 abline(h=0.5,col=2,lty=2)
}
