% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{autocorr}
\alias{autocorr}
\title{autocorr}
\usage{
autocorr(data, plot = F)
}
\arguments{
\item{data}{a list containing 1) x, the functions x_i(t) and 2) grid, the grid of time points.}

\item{plot}{a logical value. If it is TRUE, an image (heat map) of the
autocorrelation matrix is plotted. (optional)}
}
\value{
a symmetric matrix.
}
\description{
compute the autocorrelation of the sample ( x_i(t) )_{i=1,...,n}.
}
\examples{
library(RColorBrewer)
### Autocorrelation of the function x_i(t)
param <- list(n=50,p=100,beta_type="smooth")
data <- sim(param)
res_autocorr <- autocorr(data)
cols <- rev(colorRampPalette(brewer.pal(9,"YlOrRd"))(50))
image(res_autocorr,col=cols)
\donttest{
### Autocorrelation of the function beta_j(t).
data(data1)
data(param1)
res_Bliss_mult <- Bliss_multiple(data=data1,param=param1)
beta_functions_autocorr <- autocorr(list(grid = data1$grids[[1]],
                                         x = res_Bliss_mult$beta_functions[[1]]))
image(beta_functions_autocorr)
}
}
