% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bliss.R
\docType{package}
\name{bliss}
\alias{bliss}
\title{bliss: Bayesian functional Linear regression with Sparse Step functions}
\description{
A method for the Bayesian Functional Linear Regression model (functions-on-scalar),
including two estimators of the coefficient function and an estimator of its support.
A representation of the posterior distribution is also available.
}
