% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_and_graphics_functions.R
\name{dposterior}
\alias{dposterior}
\title{dposterior}
\usage{
dposterior(posterior_sample, data, Q, theta = NULL)
}
\arguments{
\item{posterior_sample}{a list given by the "Bliss_Gibbs_Sampler" function.}

\item{data}{a list containing
\describe{
\item{x}{a list of matrices, the functions x_qi(t) observed at time points given by grids,}
\item{y}{a numerical vector, the outcome values y_i.}
}}

\item{Q}{an integer, the number of covariates}

\item{theta}{a matrix or a vector which contains the parameter set. XXXX}
}
\value{
Return the (log) posterior density, the (log) likelihood and the
        (log) prior density for the given parameter set.
}
\description{
Compute the posterior density for a given parameter set. XXX (non normalisé)
}
\details{
If the option theta is NULL, the posterior density is computed for
         the MCMC sample given in the "res.Gibbs_Sampler" object.
}
\examples{
data(data1)
data(param1)
# result of res_bliss1<-fit_Bliss(data=data1,param=param1)
data(res_bliss1)
# Compute the posterior density of the MCMC sample :
res_poste <- dposterior(res_bliss1$posterior_sample,data1,Q=2)
}
