% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_and_graphics_functions.R
\name{dposterior}
\alias{dposterior}
\title{dposterior}
\usage{
dposterior(posterior_sample, data, Q, theta = NULL)
}
\arguments{
\item{posterior_sample}{a list given by the "Bliss_Gibbs_Sampler" function.}

\item{data}{a list containing
\describe{
\item{x}{a list of matrices, the functions x_qi(t) observed at time points given by grids,}
\item{y}{a numerical vector, the outcome values y_i.}
}}

\item{Q}{an integer, the number of covariates}

\item{theta}{a matrix or a vector which contains the parameter set. XXXX}
}
\value{
Return the (log) posterior density, the (log) likelihood and the
        (log) prior density for the given parameter set.
}
\description{
Compute the posterior density for a given parameter set. XXX (non normalisé)
}
\details{
If the option theta is NULL, the posterior density is computed for
         the MCMC sample given in the "res.Gibbs_Sampler" object.
}
\examples{
\donttest{
data(data1)
data(param1)
res_Bliss_mult <- fit_Bliss(data=data1,param=param1)
# Compute the posterior density of the MCMC sample :
res_poste <- dposterior(res_Bliss_mult$res.Gibbs_Sampler,data1)
### Compute the posterior density of the MCMC sample :
theta <- list( beta= rbind(c(3,4,-1), c(3,2,1)),
m = rbind(c(10,50,80),c(50,10,10)),
l = rbind(c(10,10,5),c(10,5,5)),
mu=c(1,1.5),sigma_sq=c(1,0.5))
dposterior(res_Bliss_mult$res.Gibbs_Sampler,data1,theta=theta)
}
}
