% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{\%between\%}
\alias{\%between\%}
\title{between}
\usage{
value \%between\% interval
}
\arguments{
\item{value}{a numerical value.}

\item{interval}{a numerical vector: (lower,upper).}
}
\value{
a logical value.
}
\description{
Check if a number belong to a given interval.
}
\examples{
1 \%between\% c(0,2)
2 \%between\% c(0,2)
3 \%between\% c(0,2)
}
