% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_Bliss.R
\name{image_Bliss}
\alias{image_Bliss}
\title{image_Bliss}
\usage{
image_Bliss(posterior_density_estimate, param)
}
\arguments{
\item{posterior_density_estimate}{a list. The result of the function
density_estimation.}

\item{param}{a list containing
\describe{
\item{cols}{a vector of colors for the function image (optional)}
\item{col_scale}{a character vector}
\item{ylim}{a numerical two-vector (optional)}
\item{main}{a character string.}
}}
}
\description{
plot the representation of the estimated posterior density.
}
\examples{
library(RColorBrewer)
data(data1)
data(param1)
res_Bliss_mult <- Bliss_multiple(data=data1,param=param1,density=TRUE)
param1$cols <- colorRampPalette(brewer.pal(9,"Reds"))(1e2)
image_Bliss(res_Bliss_mult$posterior_density_estimate[[1]],param1)
lines(param1$grids[[1]],res_Bliss_mult$Bliss_estimate[[1]],type="s",lwd=2)
lines(param1$grids[[1]],data1$beta_function[[1]],col=3,lwd=2,type="s")

param1$cols <- colorRampPalette(brewer.pal(9,"YlOrRd"))(1e2)
image_Bliss(res_Bliss_mult$posterior_density_estimate[[1]],param1)
lines(param1$grids[[1]],res_Bliss_mult$Bliss_estimate[[1]],type="s",lwd=2)
lines(param1$grids[[1]],data1$beta_function[[1]],col=3,lwd=2,type="s")

param1$cols <- rev(heat.colors(12))
param1$col_scale <- "quantile"
image_Bliss(res_Bliss_mult$posterior_density_estimate[[1]],param1)
lines(param1$grids[[1]],res_Bliss_mult$Bliss_estimate[[1]],type="s",lwd=2)
lines(param1$grids[[1]],data1$beta_function[[1]],col=3,lwd=2,type="s")

param1$cols <- rev(terrain.colors(12))
image_Bliss(res_Bliss_mult$posterior_density_estimate[[1]],param1)
lines(param1$grids[[1]],res_Bliss_mult$Bliss_estimate[[1]],type="s",lwd=2)
lines(param1$grids[[1]],data1$beta_function[[1]],col=2,lwd=2,type="s")

param1$cols <- rev(topo.colors(12))
image_Bliss(res_Bliss_mult$posterior_density_estimate[[1]],param1)
lines(param1$grids[[1]],res_Bliss_mult$Bliss_estimate[[1]],type="s",lwd=2)
lines(param1$grids[[1]],data1$beta_function[[1]],col=2,lwd=2,type="s")
}
