% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{integrate_trapeze}
\alias{integrate_trapeze}
\title{integrate_trapeze XXXXXXXXXXX}
\usage{
integrate_trapeze(x, y)
}
\arguments{
\item{x}{a numeric}

\item{y}{a numeric}
}
\value{
a numeric
}
\description{
integrate_trapeze
}
\examples{
# not run
}
