% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{integrate_trapeze}
\alias{integrate_trapeze}
\title{integrate_trapeze XXXXXXXXXXX
integrate_trapeze}
\usage{
integrate_trapeze(x, y)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
a graphic
}
\description{
integrate_trapeze XXXXXXXXXXX
integrate_trapeze
}
\examples{
data(data1)
data(param1)
}
