% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{interval_detection}
\alias{interval_detection}
\title{interval_detection XXXXXXXXX}
\usage{
interval_detection(beta_sample)
}
\arguments{
\item{beta_sample}{a numerical vector.}
}
\value{
a matrix with 3 columns : "begin", "end" and "value". The two first
          columns define the begin and the end of the pieces and the third
          gives the mean values of each piece.
          Or something else.
}
\description{
Determine for which intervals a functions ???
}
\examples{
data(data1)
data(param1)
# result of res_bliss1<-fit_Bliss(data=data1,param=param1)
data(res_bliss1)
intervals <- interval_detection(res_bliss1$Bliss_estimate[[1]])
\donttest{
par(mfrow=c(2,1))
plot(data1$grids[[1]],res_bliss1$Bliss_estimate[[1]],type="s")
for(k in 1:nrow(intervals)){
  segments(intervals[k,1],intervals[k,3],
          intervals[k,2],intervals[k,3],col=2,lwd=2)
}
par(mfrow=c(1,1))
}
}
