% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_and_graphics_functions.R
\name{lines_bliss}
\alias{lines_bliss}
\title{lines_bliss}
\usage{
lines_bliss(extended_grid, fct, bound = FALSE, ...)
}
\arguments{
\item{extended_grid}{a numerical vector, a new grid to provide a suitable
representation.}

\item{fct}{a numerical vector, the function to plot.}

\item{bound}{a logical value. If bound is TRUE, the plot of fct is one line.
Otherwise, several lines are used to plot fct.}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}
}
\description{
A suitable representation of the Bliss estimate.
}
\examples{
### Plot the BLiss estimate on a suitable grid
}
