% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary_and_graphics_functions.R
\name{lines_bliss}
\alias{lines_bliss}
\title{lines_bliss}
\usage{
lines_bliss(x, y, col = "black", lty = "solid")
}
\arguments{
\item{x}{the coordinates of points in the plot.}

\item{y}{the y coordinates of points in the plot.}

\item{col}{a color.}

\item{lty}{option corresponding to "linetype" of \code{geom_line}.}
}
\description{
Add a line to a plot obtained with \code{image_Bliss}.
}
\examples{
data(data1)
data(param1)
data(res_bliss1)

image_Bliss(res_bliss1$beta_posterior_density,param1,q=1) +
lines_bliss(res_bliss1$data$grids[[1]],res_bliss1$smooth_estimate[[1]])+
lines_bliss(res_bliss1$data$grids[[1]],res_bliss1$Bliss_estimate[[1]],col="purple")

}
