% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{param1}
\alias{param1}
\title{A list of param for bliss model}
\format{
a list of param for bliss model
\describe{
  \item{Q}{the number of functional covariates}
  \item{n}{the sample size}
  \item{p}{the number of observation times}
  \item{beta_shapes}{the shapes of the coefficient functions}
  \item{grids_lim}{the range of the observation times}
  \item{grids}{the grids of the observation times}
  \item{K}{the number of intervals for the coefficient function}
}
}
\usage{
param1
}
\description{
A list of param for bliss model
}
\keyword{datasets}
