% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{pdexp}
\alias{pdexp}
\title{pdexp}
\usage{
pdexp(a, l_values)
}
\arguments{
\item{a}{a positive value, the mean of the Exponential prior.}

\item{l_values}{a numerical value, the discrete support of the parameter l.}
}
\value{
a numerical vector, which is the prability function on \code{l_values}.
}
\description{
Probability function of a discretized Exponentiel distribution.
}
\examples{
pdexp(10,seq(0,1,1))

x <- seq(0,10,le=1e3)
plot(x,dexp(x,0.5),lty=2,type="l")
lines(pdexp(0.5,1:10),type="p")
}
