% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bliss_multiple.R
\name{printbliss}
\alias{printbliss}
\title{Print a bliss Object}
\usage{
printbliss(x, ...)
}
\arguments{
\item{x}{input bliss Object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print a bliss Object
}
\examples{
data(data1)
data(param1)
res_Bliss_mult <- Bliss_multiple(data=data1,param=param1,density=TRUE)
printbliss(res_Bliss_mult)
}
