% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Data.R
\name{sigmoid}
\alias{sigmoid}
\title{sigmoid}
\usage{
sigmoid(x, asym = 1, v = 1)
}
\arguments{
\item{x}{a numerical vector, time points.}

\item{asym}{a numerical value (optional), the asymptote of the sigmoid function.}

\item{v}{a numerical value (optional), related to the slope at the origin.}
}
\value{
a numerical vector.
}
\description{
Compute a sigmoid function.
}
\details{
see the function \code{\link[=sim_x]{sim_x}}.
}
\examples{
## Test 1 :
x <- seq(-7,7,0.1)
y <- sigmoid(x)
plot(x,y,type="l",main="Sigmoid function")
## Test 2 :
x  <- seq(-7,7,0.1)
y  <- sigmoid(x)
y2 <- sigmoid(x,asym=0.5)
y3 <- sigmoid(x,v   =  5)
plot(x,y,type="l",main="Other sigmoid functions")
lines(x,y2,col=2)
lines(x,y3,col=3)
}
