% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Data.R
\name{sim}
\alias{sim}
\title{sim}
\usage{
sim(param, progress = FALSE)
}
\arguments{
\item{param}{a list containing:
\describe{
 \item{Q}{an integer, the number of functional covariates.}
 \item{n}{an integer, the sample size.}
 \item{p}{a vector of integers, the qth component is the number of
       times for the qth covariate.}
 \item{grids}{a list of numerical vectors, the qth vector is the grid
       of time points for the qth functional covariate (optional, can be
       deduced from grid_lim and p if it is \code{NULL}).}
 \item{grids_lim}{a list of numerical vectors, the qth item is the lower
       and upper boundaries of the domain for the qth functional covariate. (optional)}
 \item{mu}{a numerical value, the intercept of the model. (optional)}
 \item{r}{a nonnegative value, the signal to noise ratio. (optional)}
 \item{link}{a function. XXXX (optional)}
 \item{beta_shapes}{a character vector. The qth item indicates the shape of
       the coefficient function associated to the qth functional covariate.}
 \item{x_shapes}{a character vector. The qth item indicates the shape of the
       functional covariate observations. (optional)}
 \item{autocorr_diag}{a list of numerical vectors, the qth vector is the
       diagonal of the autocorrelation matrix of the qth functional
       covariate (optional).}
 \item{autocorr_spread}{a vector of numerical values which are related to the
       autocorrelation of the functional covariates. (optional)}
 \item{correlation}{a matrix which gives the correlation structure between the
       functional covariates (optional).}
}}

\item{progress}{a logical value. If TRUE, the algorithm progress is displayed.
(optional)}
}
\value{
a list containing:
\describe{
 \item{Q}{an integer, the number of functional covariates.}
 \item{y}{a numerical vector, the outcome observations.}
 \item{x}{a list of matrices, the qth matrix contains the observation of the
       qth functional covariate on a grid of time points given with \code{grids}.}
 \item{grids}{a list of numerical vectors, the qth vector is the grid of
       time points for the qth functional covariate.}
 \item{betas}{a list of numerical vectors, the qth vector is the coefficient
       function associated to the qth covariate on a grid of time points
       given with \code{grids}.}
}
}
\description{
Simulate a dataset for the Function Linear Regression model.
}
\examples{
\donttest{
#### NOT RUN
library(RColorBrewer)
param <- list(Q=2,n=25,p=c(50,50),grids_lim=list(c(0,1),c(-1,2)))
data <- sim(param)
data$y
data$expe
cols <- colorRampPalette(brewer.pal(9,"YlOrRd"))(10)
q=2
par(mfrow=c(2,1))
matplot(data$grids[[q]],t(data$x_mult[[q]]),type="l",lty=1,col=cols)
plot(data$grids[[q]],data$beta_function_mult[[q]],type="l")
abline(h=0,lty=2,col="gray")
par(mfrow=c(1,1))
}
}
