% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{support_estimation}
\alias{support_estimation}
\title{support_estimation XXXXXXXXX}
\usage{
support_estimation(beta_sample, gamma = 0.5)
}
\arguments{
\item{beta_sample}{a matrix. Each row is a coefficient function computed from the
posterior sample.}

\item{gamma}{a numeric, default 0.5}
}
\value{
a list containing:
\describe{
 \item{alpha}{a numerical vector. XXXXXXX}
 \item{estimate}{a numerical vector. XXXXXXX}
}
}
\description{
Compute the support estimate.
}
\examples{
\donttest{
data(data1)
data(param1)
res_Bliss_mult <- fit_Bliss(data=data1,param=param1)
res_support <- support_estimation(res_Bliss_mult$beta_sample[[1]])
 ### The estimate
 res_support$estimate
 ### Plot the result
 grid <- res_Bliss_mult$param$grids[[1]]
 plot(grid,res_support$alpha_t,ylim=c(0,1),type="l",xlab="",ylab="")
 for(k in 1:nrow(res_support$estimate)){
    segments(grid[res_support$estimate[k,1]],0.5,
           grid[res_support$estimate[k,2]],0.5,lwd=2,col=2)
   points(grid[res_support$estimate[k,1]],0.5,pch="|",lwd=2,col=2)
   points(grid[res_support$estimate[k,2]],0.5,pch="|",lwd=2,col=2)
 }
 abline(h=0.5,col=2,lty=2)
}
}
