% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{support_estimation}
\alias{support_estimation}
\title{support_estimation}
\usage{
support_estimation(beta_sample, param)
}
\arguments{
\item{beta_sample}{the result of the function \code{compute_beta_sample}.}

\item{param}{a list containing the value \code{Q} and an optional
parameter \code{gamma}.}
}
\value{
a list containing:
\describe{
 \item{alpha}{a numerical vector. The approximated posterior probabilities
       that the coefficient function support covers \code{t} for each time
       points \code{t}.}
 \item{estimate}{a numerical vector, the support estimate.}
 \item{estimate_fct}{a numerical vector, another version of the support
       estimate.}
}
}
\description{
Compute the support estimate.
}
\examples{
data(data1)
data(param1)
data(res_bliss1)
param1$Q <- length(data1$x)

res_support <- support_estimation(res_bliss1$beta_sample,param1)
}
